+++ internal/clients/fleet/fleet.go
@@ -8,8 +8,7 @@ import (
 	"net/http"
 
 	fleetapi "github.com/elastic/terraform-provider-elasticstack/generated/fleet"
-	fwdiag "github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
 )
 
 var (
@@ -17,7 +16,7 @@ var (
 )
 
 // AllEnrollmentTokens reads all enrollment tokens from the API.
-func AllEnrollmentTokens(ctx context.Context, client *Client) ([]fleetapi.EnrollmentApiKey, fwdiag.Diagnostics) {
+func AllEnrollmentTokens(ctx context.Context, client *Client) ([]fleetapi.EnrollmentApiKey, diag.Diagnostics) {
 	resp, err := client.API.GetEnrollmentApiKeysWithResponse(ctx)
 	if err != nil {
 		return nil, fromErr(err)
@@ -26,11 +25,11 @@ func AllEnrollmentTokens(ctx context.Context, client *Client) ([]fleetapi.Enroll
 	if resp.StatusCode() == http.StatusOK {
 		return resp.JSON200.Items, nil
 	}
-	return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+	return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 }
 
 // GetEnrollmentTokensByPolicy Get enrollment tokens by given policy ID
-func GetEnrollmentTokensByPolicy(ctx context.Context, client *Client, policyID string) ([]fleetapi.EnrollmentApiKey, fwdiag.Diagnostics) {
+func GetEnrollmentTokensByPolicy(ctx context.Context, client *Client, policyID string) ([]fleetapi.EnrollmentApiKey, diag.Diagnostics) {
 	resp, err := client.API.GetEnrollmentApiKeysWithResponse(ctx, func(ctx context.Context, req *http.Request) error {
 		q := req.URL.Query()
 		q.Set("kuery", "policy_id:"+policyID)
@@ -45,11 +44,11 @@ func GetEnrollmentTokensByPolicy(ctx context.Context, client *Client, policyID s
 	if resp.StatusCode() == http.StatusOK {
 		return resp.JSON200.Items, nil
 	}
-	return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+	return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 }
 
 // ReadAgentPolicy reads a specific agent policy from the API.
-func ReadAgentPolicy(ctx context.Context, client *Client, id string) (*fleetapi.AgentPolicy, fwdiag.Diagnostics) {
+func ReadAgentPolicy(ctx context.Context, client *Client, id string) (*fleetapi.AgentPolicy, diag.Diagnostics) {
 	resp, err := client.API.AgentPolicyInfoWithResponse(ctx, id)
 	if err != nil {
 		return nil, fromErr(err)
@@ -61,12 +60,12 @@ func ReadAgentPolicy(ctx context.Context, client *Client, id string) (*fleetapi.
 	case http.StatusNotFound:
 		return nil, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // CreateAgentPolicy creates a new agent policy.
-func CreateAgentPolicy(ctx context.Context, client *Client, req fleetapi.AgentPolicyCreateRequest) (*fleetapi.AgentPolicy, fwdiag.Diagnostics) {
+func CreateAgentPolicy(ctx context.Context, client *Client, req fleetapi.AgentPolicyCreateRequest) (*fleetapi.AgentPolicy, diag.Diagnostics) {
 	resp, err := client.API.CreateAgentPolicyWithResponse(ctx, req)
 	if err != nil {
 		return nil, fromErr(err)
@@ -76,12 +75,12 @@ func CreateAgentPolicy(ctx context.Context, client *Client, req fleetapi.AgentPo
 	case http.StatusOK:
 		return resp.JSON200.Item, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // UpdateAgentPolicy updates an existing agent policy.
-func UpdateAgentPolicy(ctx context.Context, client *Client, id string, req fleetapi.AgentPolicyUpdateRequest) (*fleetapi.AgentPolicy, fwdiag.Diagnostics) {
+func UpdateAgentPolicy(ctx context.Context, client *Client, id string, req fleetapi.AgentPolicyUpdateRequest) (*fleetapi.AgentPolicy, diag.Diagnostics) {
 	resp, err := client.API.UpdateAgentPolicyWithResponse(ctx, id, req)
 	if err != nil {
 		return nil, fromErr(err)
@@ -91,12 +90,12 @@ func UpdateAgentPolicy(ctx context.Context, client *Client, id string, req fleet
 	case http.StatusOK:
 		return &resp.JSON200.Item, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // DeleteAgentPolicy deletes an existing agent policy
-func DeleteAgentPolicy(ctx context.Context, client *Client, id string) fwdiag.Diagnostics {
+func DeleteAgentPolicy(ctx context.Context, client *Client, id string) diag.Diagnostics {
 	body := fleetapi.DeleteAgentPolicyJSONRequestBody{
 		AgentPolicyId: id,
 	}
@@ -112,7 +111,7 @@ func DeleteAgentPolicy(ctx context.Context, client *Client, id string) fwdiag.Di
 	case http.StatusNotFound:
 		return nil
 	default:
-		return reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
@@ -120,7 +119,7 @@ func DeleteAgentPolicy(ctx context.Context, client *Client, id string) fwdiag.Di
 func ReadOutput(ctx context.Context, client *Client, id string) (*fleetapi.OutputCreateRequest, diag.Diagnostics) {
 	resp, err := client.API.GetOutputWithResponse(ctx, id)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		return nil, fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -137,7 +136,7 @@ func ReadOutput(ctx context.Context, client *Client, id string) (*fleetapi.Outpu
 func CreateOutput(ctx context.Context, client *Client, req fleetapi.PostOutputsJSONRequestBody) (*fleetapi.OutputCreateRequest, diag.Diagnostics) {
 	resp, err := client.API.PostOutputsWithResponse(ctx, req)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		return nil, fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -152,7 +151,7 @@ func CreateOutput(ctx context.Context, client *Client, req fleetapi.PostOutputsJ
 func UpdateOutput(ctx context.Context, client *Client, id string, req fleetapi.UpdateOutputJSONRequestBody) (*fleetapi.OutputUpdateRequest, diag.Diagnostics) {
 	resp, err := client.API.UpdateOutputWithResponse(ctx, id, req)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		return nil, fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -167,7 +166,7 @@ func UpdateOutput(ctx context.Context, client *Client, id string, req fleetapi.U
 func DeleteOutput(ctx context.Context, client *Client, id string) diag.Diagnostics {
 	resp, err := client.API.DeleteOutputWithResponse(ctx, id)
 	if err != nil {
-		return diag.FromErr(err)
+		return fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -181,7 +180,7 @@ func DeleteOutput(ctx context.Context, client *Client, id string) diag.Diagnosti
 }
 
 // ReadFleetServerHost reads a specific fleet server host from the API.
-func ReadFleetServerHost(ctx context.Context, client *Client, id string) (*fleetapi.FleetServerHost, fwdiag.Diagnostics) {
+func ReadFleetServerHost(ctx context.Context, client *Client, id string) (*fleetapi.FleetServerHost, diag.Diagnostics) {
 	resp, err := client.API.GetOneFleetServerHostsWithResponse(ctx, id)
 	if err != nil {
 		return nil, fromErr(err)
@@ -193,12 +192,12 @@ func ReadFleetServerHost(ctx context.Context, client *Client, id string) (*fleet
 	case http.StatusNotFound:
 		return nil, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // CreateFleetServerHost creates a new fleet server host.
-func CreateFleetServerHost(ctx context.Context, client *Client, req fleetapi.PostFleetServerHostsJSONRequestBody) (*fleetapi.FleetServerHost, fwdiag.Diagnostics) {
+func CreateFleetServerHost(ctx context.Context, client *Client, req fleetapi.PostFleetServerHostsJSONRequestBody) (*fleetapi.FleetServerHost, diag.Diagnostics) {
 	resp, err := client.API.PostFleetServerHostsWithResponse(ctx, req)
 	if err != nil {
 		return nil, fromErr(err)
@@ -208,12 +207,12 @@ func CreateFleetServerHost(ctx context.Context, client *Client, req fleetapi.Pos
 	case http.StatusOK:
 		return resp.JSON200.Item, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // UpdateFleetServerHost updates an existing fleet server host.
-func UpdateFleetServerHost(ctx context.Context, client *Client, id string, req fleetapi.UpdateFleetServerHostsJSONRequestBody) (*fleetapi.FleetServerHost, fwdiag.Diagnostics) {
+func UpdateFleetServerHost(ctx context.Context, client *Client, id string, req fleetapi.UpdateFleetServerHostsJSONRequestBody) (*fleetapi.FleetServerHost, diag.Diagnostics) {
 	resp, err := client.API.UpdateFleetServerHostsWithResponse(ctx, id, req)
 	if err != nil {
 		return nil, fromErr(err)
@@ -223,12 +222,12 @@ func UpdateFleetServerHost(ctx context.Context, client *Client, id string, req f
 	case http.StatusOK:
 		return &resp.JSON200.Item, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // DeleteFleetServerHost deletes an existing fleet server host.
-func DeleteFleetServerHost(ctx context.Context, client *Client, id string) fwdiag.Diagnostics {
+func DeleteFleetServerHost(ctx context.Context, client *Client, id string) diag.Diagnostics {
 	resp, err := client.API.DeleteFleetServerHostsWithResponse(ctx, id)
 	if err != nil {
 		return fromErr(err)
@@ -240,7 +239,7 @@ func DeleteFleetServerHost(ctx context.Context, client *Client, id string) fwdia
 	case http.StatusNotFound:
 		return nil
 	default:
-		return reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
@@ -253,7 +252,7 @@ func ReadPackagePolicy(ctx context.Context, client *Client, id string) (*fleetap
 
 	resp, err := client.API.GetPackagePolicyWithResponse(ctx, id, &params)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		return nil, fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -275,7 +274,7 @@ func CreatePackagePolicy(ctx context.Context, client *Client, req fleetapi.Creat
 
 	resp, err := client.API.CreatePackagePolicyWithResponse(ctx, &params, req)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		return nil, fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -295,7 +294,7 @@ func UpdatePackagePolicy(ctx context.Context, client *Client, id string, req fle
 
 	resp, err := client.API.UpdatePackagePolicyWithResponse(ctx, id, &params, req)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		return nil, fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -311,7 +310,7 @@ func DeletePackagePolicy(ctx context.Context, client *Client, id string, force b
 	params := fleetapi.DeletePackagePolicyParams{Force: &force}
 	resp, err := client.API.DeletePackagePolicyWithResponse(ctx, id, &params)
 	if err != nil {
-		return diag.FromErr(err)
+		return fromErr(err)
 	}
 
 	switch resp.StatusCode() {
@@ -325,7 +324,7 @@ func DeletePackagePolicy(ctx context.Context, client *Client, id string, force b
 }
 
 // ReadPackage reads a specific package from the API.
-func ReadPackage(ctx context.Context, client *Client, name, version string) fwdiag.Diagnostics {
+func ReadPackage(ctx context.Context, client *Client, name, version string) diag.Diagnostics {
 	params := fleetapi.GetPackageParams{}
 
 	resp, err := client.API.GetPackage(ctx, name, version, &params)
@@ -345,12 +344,12 @@ func ReadPackage(ctx context.Context, client *Client, name, version string) fwdi
 			return fromErr(err)
 		}
 
-		return reportUnknownErrorFw(resp.StatusCode, errData)
+		return reportUnknownError(resp.StatusCode, errData)
 	}
 }
 
 // InstallPackage installs a package.
-func InstallPackage(ctx context.Context, client *Client, name, version string, force bool) fwdiag.Diagnostics {
+func InstallPackage(ctx context.Context, client *Client, name, version string, force bool) diag.Diagnostics {
 	params := fleetapi.InstallPackageParams{}
 	body := fleetapi.InstallPackageJSONRequestBody{
 		Force:             &force,
@@ -372,12 +371,12 @@ func InstallPackage(ctx context.Context, client *Client, name, version string, f
 			return fromErr(err)
 		}
 
-		return reportUnknownErrorFw(resp.StatusCode, errData)
+		return reportUnknownError(resp.StatusCode, errData)
 	}
 }
 
 // Uninstall uninstalls a package.
-func Uninstall(ctx context.Context, client *Client, name, version string, force bool) fwdiag.Diagnostics {
+func Uninstall(ctx context.Context, client *Client, name, version string, force bool) diag.Diagnostics {
 	params := fleetapi.DeletePackageParams{}
 	body := fleetapi.DeletePackageJSONRequestBody{
 		Force: &force,
@@ -394,12 +393,12 @@ func Uninstall(ctx context.Context, client *Client, name, version string, force 
 	case http.StatusNotFound:
 		return nil
 	default:
-		return reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // AllPackages returns information about the latest packages known to Fleet.
-func AllPackages(ctx context.Context, client *Client, prerelease bool) ([]fleetapi.SearchResult, fwdiag.Diagnostics) {
+func AllPackages(ctx context.Context, client *Client, prerelease bool) ([]fleetapi.SearchResult, diag.Diagnostics) {
 	params := fleetapi.ListAllPackagesParams{
 		Prerelease: &prerelease,
 	}
@@ -413,33 +412,23 @@ func AllPackages(ctx context.Context, client *Client, prerelease bool) ([]fleeta
 	case http.StatusOK:
 		return resp.JSON200.Items, nil
 	default:
-		return nil, reportUnknownErrorFw(resp.StatusCode(), resp.Body)
+		return nil, reportUnknownError(resp.StatusCode(), resp.Body)
 	}
 }
 
 // fromErr recreates the sdkdiag.FromErr functionality.
-func fromErr(err error) fwdiag.Diagnostics {
+func fromErr(err error) diag.Diagnostics {
 	if err == nil {
 		return nil
 	}
-	return fwdiag.Diagnostics{
-		fwdiag.NewErrorDiagnostic(err.Error(), ""),
+	return diag.Diagnostics{
+		diag.NewErrorDiagnostic(err.Error(), ""),
 	}
 }
 
 func reportUnknownError(statusCode int, body []byte) diag.Diagnostics {
 	return diag.Diagnostics{
-		diag.Diagnostic{
-			Severity: diag.Error,
-			Summary:  fmt.Sprintf("Unexpected status code from server: got HTTP %d", statusCode),
-			Detail:   string(body),
-		},
-	}
-}
-
-func reportUnknownErrorFw(statusCode int, body []byte) fwdiag.Diagnostics {
-	return fwdiag.Diagnostics{
-		fwdiag.NewErrorDiagnostic(
+		diag.NewErrorDiagnostic(
 			fmt.Sprintf("Unexpected status code from server: got HTTP %d", statusCode),
 			string(body),
 		),