+++ internal/fleet/integration_policy/private.go
@@ -1,0 +1,34 @@
+package integration_policy
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+)
+
+// Equivalent to privatestate.ProviderData
+type privateData interface {
+	// GetKey returns the private state data associated with the given key.
+	//
+	// If the key is reserved for framework usage, an error diagnostic
+	// is returned. If the key is valid, but private state data is not found,
+	// nil is returned.
+	//
+	// The naming of keys only matters in context of a single resource,
+	// however care should be taken that any historical keys are not reused
+	// without accounting for older resource instances that may still have
+	// older data at the key.
+	GetKey(ctx context.Context, key string) ([]byte, diag.Diagnostics)
+
+	// SetKey sets the private state data at the given key.
+	//
+	// If the key is reserved for framework usage, an error diagnostic
+	// is returned. The data must be valid JSON and UTF-8 safe or an error
+	// diagnostic is returned.
+	//
+	// The naming of keys only matters in context of a single resource,
+	// however care should be taken that any historical keys are not reused
+	// without accounting for older resource instances that may still have
+	// older data at the key.
+	SetKey(ctx context.Context, key string, value []byte) diag.Diagnostics
+}