+++ internal/fleet/output/resource_test.go
@@ -1,14 +1,14 @@
-package fleet_test
+package output_test
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/elastic/terraform-provider-elasticstack/internal/acctest"
 	"github.com/elastic/terraform-provider-elasticstack/internal/clients"
 	"github.com/elastic/terraform-provider-elasticstack/internal/clients/fleet"
+	"github.com/elastic/terraform-provider-elasticstack/internal/utils"
 	"github.com/elastic/terraform-provider-elasticstack/internal/versionutils"
 	"github.com/hashicorp/go-version"
 	sdkacctest "github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
@@ -19,7 +19,7 @@ import (
 var minVersionOutput = version.Must(version.NewVersion("8.6.0"))
 
 func TestAccResourceOutputElasticsearch(t *testing.T) {
-	policyName := sdkacctest.RandStringFromCharSet(22, sdkacctest.CharSetAlphaNum)
+	policyName := sdkacctest.RandString(22)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
@@ -31,7 +31,7 @@ func TestAccResourceOutputElasticsearch(t *testing.T) {
 				Config:   testAccResourceOutputCreateElasticsearch(policyName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "name", fmt.Sprintf("Elasticsearch Output %s", policyName)),
-					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", "elasticsearch-output"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", fmt.Sprintf("%s-elasticsearch-output", policyName)),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "type", "elasticsearch"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "config_yaml", "\"ssl.verification_mode\": \"none\"\n"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "default_integrations", "false"),
@@ -44,7 +44,7 @@ func TestAccResourceOutputElasticsearch(t *testing.T) {
 				Config:   testAccResourceOutputUpdateElasticsearch(policyName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "name", fmt.Sprintf("Updated Elasticsearch Output %s", policyName)),
-					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", "elasticsearch-output"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", fmt.Sprintf("%s-elasticsearch-output", policyName)),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "type", "elasticsearch"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "config_yaml", "\"ssl.verification_mode\": \"none\"\n"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "default_integrations", "false"),
@@ -54,6 +54,7 @@ func TestAccResourceOutputElasticsearch(t *testing.T) {
 			},
 			{
 				SkipFunc:          versionutils.CheckIfVersionIsUnsupported(minVersionOutput),
+				Config:            testAccResourceOutputUpdateElasticsearch(policyName),
 				ResourceName:      "elasticstack_fleet_output.test_output",
 				ImportState:       true,
 				ImportStateVerify: true,
@@ -63,7 +64,7 @@ func TestAccResourceOutputElasticsearch(t *testing.T) {
 }
 
 func TestAccResourceOutputLogstash(t *testing.T) {
-	policyName := sdkacctest.RandStringFromCharSet(22, sdkacctest.CharSetAlphaNum)
+	policyName := sdkacctest.RandString(22)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
@@ -75,7 +76,7 @@ func TestAccResourceOutputLogstash(t *testing.T) {
 				Config:   testAccResourceOutputCreateLogstash(policyName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "name", fmt.Sprintf("Logstash Output %s", policyName)),
-					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", "logstash-output"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", fmt.Sprintf("%s-logstash-output", policyName)),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "type", "logstash"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "config_yaml", "\"ssl.verification_mode\": \"none\"\n"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "default_integrations", "false"),
@@ -91,7 +92,7 @@ func TestAccResourceOutputLogstash(t *testing.T) {
 				Config:   testAccResourceOutputLogstashUpdate(policyName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "name", fmt.Sprintf("Updated Logstash Output %s", policyName)),
-					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", "logstash-output"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "id", fmt.Sprintf("%s-logstash-output", policyName)),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "type", "logstash"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "config_yaml", "\"ssl.verification_mode\": \"none\"\n"),
 					resource.TestCheckResourceAttr("elasticstack_fleet_output.test_output", "default_integrations", "false"),
@@ -114,8 +115,8 @@ provider "elasticstack" {
 }
 
 resource "elasticstack_fleet_output" "test_output" {
-  name                 = "%s"
-  output_id            = "elasticsearch-output"
+  name                 = "Elasticsearch Output %s"
+  output_id            = "%s-elasticsearch-output"
   type                 = "elasticsearch"
   config_yaml = yamlencode({
     "ssl.verification_mode" : "none"
@@ -126,7 +127,7 @@ resource "elasticstack_fleet_output" "test_output" {
     "https://elasticsearch:9200"
   ]
 }
-`, fmt.Sprintf("Elasticsearch Output %s", id))
+`, id, id)
 }
 
 func testAccResourceOutputUpdateElasticsearch(id string) string {
@@ -137,8 +138,8 @@ provider "elasticstack" {
 }
 
 resource "elasticstack_fleet_output" "test_output" {
-  name                 = "%s"
-  output_id            = "elasticsearch-output"
+  name                 = "Updated Elasticsearch Output %s"
+  output_id            = "%s-elasticsearch-output"
   type                 = "elasticsearch"
   config_yaml = yamlencode({
     "ssl.verification_mode" : "none"
@@ -149,8 +150,7 @@ resource "elasticstack_fleet_output" "test_output" {
     "https://elasticsearch:9200"
   ]
 }
-
-`, fmt.Sprintf("Updated Elasticsearch Output %s", id))
+`, id, id)
 }
 
 func testAccResourceOutputCreateLogstash(id string) string {
@@ -161,9 +161,9 @@ provider "elasticstack" {
 }
 
 resource "elasticstack_fleet_output" "test_output" {
-  name                 = "%s"
+  name                 = "Logstash Output %s"
   type                 = "logstash"
-  output_id            = "logstash-output"
+  output_id            = "%s-logstash-output"
   config_yaml = yamlencode({
     "ssl.verification_mode" : "none"
   })
@@ -178,7 +178,7 @@ resource "elasticstack_fleet_output" "test_output" {
 	key                     = "placeholder"
   }
 }
-`, fmt.Sprintf("Logstash Output %s", id))
+`, id, id)
 }
 
 func testAccResourceOutputLogstashUpdate(id string) string {
@@ -189,9 +189,9 @@ provider "elasticstack" {
 }
 
 resource "elasticstack_fleet_output" "test_output" {
-  name                 = "%s"
+  name                 = "Updated Logstash Output %s"
+  output_id            = "%s-logstash-output"
   type                 = "logstash"
-  output_id            = "logstash-output"
   config_yaml = yamlencode({
     "ssl.verification_mode" : "none"
   })
@@ -206,8 +206,7 @@ resource "elasticstack_fleet_output" "test_output" {
 	key                     = "placeholder"
   }
 }
-
-`, fmt.Sprintf("Updated Logstash Output %s", id))
+`, id, id)
 }
 
 func checkResourceOutputDestroy(s *terraform.State) error {
@@ -225,11 +224,11 @@ func checkResourceOutputDestroy(s *terraform.State) error {
 		if err != nil {
 			return err
 		}
-		packagePolicy, diag := fleet.ReadOutput(context.Background(), fleetClient, rs.Primary.ID)
-		if diag.HasError() {
-			return errors.New(diag[0].Summary)
+		output, diags := fleet.ReadOutput(context.Background(), fleetClient, rs.Primary.ID)
+		if diags.HasError() {
+			return utils.FwDiagsAsError(diags)
 		}
-		if packagePolicy != nil {
+		if output != nil {
 			return fmt.Errorf("output id=%v still exists, but it should have been removed", rs.Primary.ID)
 		}
 	}